# Python: bindings optimizations and support numpy arrays for xfer_dlist