Speedup for 1M descriptor lists:

get_xfer_descs with numpy input: x3.7 vs list (82ms -> 22ms)
get_xfer_descs with list input: x1.3 (82ms -> 63ms)
get_reg_descs with numpy input: x2.7 vs get_reg_descs (146ms -> 54ms)
get_reg_descs with list input: x1.2 (146ms -> 122ms)
make_prepped_xfer with list: x1 (74ms -> 74ms)
make_prepped_xfer with numpy in: x1.7 vs list (74ms -> 44ms)